requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0

# Database - Compatible versions
supabase>=2.11.0
postgrest>=0.18.0

# Web framework
flask==2.3.3
werkzeug==2.3.7
jinja2==3.1.2
gunicorn==21.2.0

# Monitoring and security
psutil==5.9.6
sentry-sdk==1.38.0
prometheus-client==0.19.0
cryptography==41.0.7

# Optional (can be commented out if not needed)
redis==5.0.1
httpx>=0.24
